Method extraction and refactoring user entity
