+++ rest/src/main/resources/application.properties
@@ -9,7 +9,8 @@ spring.datasource.url=jdbc:postgresql://localhost:5432/events_database?createDat
 spring.datasource.username=postgres
 spring.datasource.password=111111
 spring.datasource.driver-class-name=org.postgresql.Driver
-spring.jpa.database-platform=org.hibernate.dialect.PostgreSQLDialect
 spring.jpa.hibernate.ddl-auto=create-drop
 
 application.security.jwt.secret-key=ae131faefd2452cade07273f79bc0555b7f6fe1f8b4773840e741b34864242cb
+application.security.jwt..expiration=86400000
+application.security.jwt.refresh-token.expiration=604800000
